Function mmal_sys::mmal_port_connect

pub unsafe extern "C" fn mmal_port_connect(
    port: *mut MMAL_PORT_T, 
    other_port: *mut MMAL_PORT_T
) -> Type

Connect an output port to an input port.

When connected and enabled, buffers will automatically progress from the output port to the input port when they become available, and released back to the output port when no longer required by the input port.

Ports can be given either way around, but one must be an output port and the other must be an input port. Neither can be connected or enabled already. The format of the output port will be applied to the input port on connection.

@param port One of the ports to connect. @param other_port The other port to connect. @return MMAL_SUCCESS on success.