Struct mmal_sys::MMAL_COMPONENT_T

#[repr(C)]

pub struct MMAL_COMPONENT_T {
    pub priv_: *mut MMAL_COMPONENT_PRIVATE_T,
    pub userdata: *mut MMAL_COMPONENT_USERDATA_T,
    pub name: *const c_char,
    pub is_enabled: u32,
    pub control: *mut MMAL_PORT_T,
    pub input_num: u32,
    pub input: *mut *mut MMAL_PORT_T,
    pub output_num: u32,
    pub output: *mut *mut MMAL_PORT_T,
    pub clock_num: u32,
    pub clock: *mut *mut MMAL_PORT_T,
    pub port_num: u32,
    pub port: *mut *mut MMAL_PORT_T,
    pub id: u32,
}

Definition of a component.

Fields
priv_: *mut MMAL_COMPONENT_PRIVATE_T

Pointer to the private data of the module in use

userdata: *mut MMAL_COMPONENT_USERDATA_T

Pointer to private data of the client

name: *const c_char

Component name

is_enabled: u32

Specifies whether the component is enabled or not

control: *mut MMAL_PORT_T

All components expose a control port. The control port is used by clients to set / get parameters that are global to the component. It is also used to receive events, which again are global to the component. To be able to receive events, the client needs to enable and register a callback on the control port.

input_num: u32

< Number of input ports

input: *mut *mut MMAL_PORT_T

< Array of input ports

output_num: u32

< Number of output ports

output: *mut *mut MMAL_PORT_T

< Array of output ports

clock_num: u32

< Number of clock ports

clock: *mut *mut MMAL_PORT_T

< Array of clock ports

port_num: u32

< Total number of ports

port: *mut *mut MMAL_PORT_T

< Array of all the ports (control/input/output/clock)

id: u32

Uniquely identifies the component's instance within the MMAL context / process. For debugging.