Struct mmal_sys::MMAL_DISPLAYREGION_T

#[repr(C)]

pub struct MMAL_DISPLAYREGION_T {
    pub hdr: MMAL_PARAMETER_HEADER_T,
    pub set: u32,
    pub display_num: u32,
    pub fullscreen: MMAL_BOOL_T,
    pub transform: MMAL_DISPLAYTRANSFORM_T,
    pub dest_rect: MMAL_RECT_T,
    pub src_rect: MMAL_RECT_T,
    pub noaspect: MMAL_BOOL_T,
    pub mode: MMAL_DISPLAYMODE_T,
    pub pixel_x: u32,
    pub pixel_y: u32,
    pub layer: i32,
    pub copyprotect_required: MMAL_BOOL_T,
    pub alpha: u32,
}

This config sets the output display device, as well as the region used on the output display, any display transformation, and some flags to indicate how to scale the image.

Fields
hdr: MMAL_PARAMETER_HEADER_T set: u32

Bitfield that indicates which fields are set and should be used. All other fields will maintain their current value. \ref MMAL_DISPLAYSET_T defines the bits that can be combined.

display_num: u32

Describes the display output device, with 0 typically being a directly connected LCD display. The actual values will depend on the hardware. Code using hard-wired numbers (e.g. 2) is certain to fail.

fullscreen: MMAL_BOOL_T

Indicates that we are using the full device screen area, rather than a window of the display. If zero, then dest_rect is used to specify a region of the display to use.

transform: MMAL_DISPLAYTRANSFORM_T

Indicates any rotation or flipping used to map frames onto the natural display orientation.

dest_rect: MMAL_RECT_T

Where to display the frame within the screen, if fullscreen is zero.

src_rect: MMAL_RECT_T

Indicates which area of the frame to display. If all values are zero, the whole frame will be used.

noaspect: MMAL_BOOL_T

If set to non-zero, indicates that any display scaling should disregard the aspect ratio of the frame region being displayed.

mode: MMAL_DISPLAYMODE_T

Indicates how the image should be scaled to fit the display. \code MMAL_DISPLAY_MODE_FILL \endcode indicates that the image should fill the screen by potentially cropping the frames. Setting \code mode \endcode to \code MMAL_DISPLAY_MODE_LETTERBOX \endcode indicates that all the source region should be displayed and black bars added if necessary.

pixel_x: u32

If non-zero, defines the width of a source pixel relative to \code pixel_y \endcode. If zero, then pixels default to being square.

pixel_y: u32

If non-zero, defines the height of a source pixel relative to \code pixel_x \endcode. If zero, then pixels default to being square.

layer: i32

Sets the relative depth of the images, with greater values being in front of smaller values.

copyprotect_required: MMAL_BOOL_T

Set to non-zero to ensure copy protection is used on output.

alpha: u32

Level of opacity of the layer, where zero is fully transparent and 255 is fully opaque.