Struct mmal_sys::VCOS_BLOCKPOOL_TAG

#[repr(C)]

pub struct VCOS_BLOCKPOOL_TAG {
    pub magic: u32,
    pub mutex: VCOS_MUTEX_T,
    pub align: VCOS_UNSIGNED,
    pub flags: VCOS_UNSIGNED,
    pub block_data_size: usize,
    pub block_size: usize,
    pub name: *const c_char,
    pub num_subpools: VCOS_UNSIGNED,
    pub num_extension_blocks: VCOS_UNSIGNED,
    pub subpools: [VCOS_BLOCKPOOL_SUBPOOL_T; 8],
}

Fields
magic: u32

VCOS_BLOCKPOOL_MAGIC

mutex: VCOS_MUTEX_T

Thread safety for Alloc, Free, Delete, Stats

align: VCOS_UNSIGNED

Alignment of block data pointers

flags: VCOS_UNSIGNED

Flags for future use e.g. cache options

block_data_size: usize

The size of the block data

block_size: usize

Block size inc overheads

name: *const c_char

Name for debugging

num_subpools: VCOS_UNSIGNED num_extension_blocks: VCOS_UNSIGNED

Number of blocks in each dynamically allocated subpool

subpools: [VCOS_BLOCKPOOL_SUBPOOL_T; 8]

Array of subpools. Subpool zero is is not deleted until the pool is destroed. If the index of the pool is < num_subpools and subpool[index.mem] is null then the subpool entry is valid but "not currently allocated"