Struct mmal_sys::VCOS_TIMER_T

#[repr(C)]

pub struct VCOS_TIMER_T {
    pub thread: pthread_t,
    pub lock: pthread_mutex_t,
    pub __bindgen_padding_0: u32,
    pub settings_changed: pthread_cond_t,
    pub quit: c_int,
    pub expires: timespec,
    pub orig_expiration_routine: Option<unsafe extern "C" fn(arg1: *mut c_void)>,
    pub orig_context: *mut c_void,
    pub __bindgen_padding_1: u32,
}

Fields
thread: pthread_t

< id of the timer thread

lock: pthread_mutex_t

< lock protecting all other members of the struct

__bindgen_padding_0: u32 settings_changed: pthread_cond_t

< cond. var. for informing the timer thread about changes

quit: c_int

< non-zero if the timer thread is requested to quit

expires: timespec

< absolute time of next expiration, or 0 if disarmed

orig_expiration_routine: Option<unsafe extern "C" fn(arg1: *mut c_void)>

< the expiration routine provided by the user of the timer

orig_context: *mut c_void

< the context for exp. routine provided by the user

__bindgen_padding_1: u32